HTML XModule substitutes %%USER_ID%%
Allows course authors to use the string %%USER_ID%% in an HTML
XModule and have it get substituted with the anonymous (pseudonymous?)
user ID. One possible use case for this is so that instructors can
embed links/iframes/etc. to third-party survey sites like Qualtrics
that take in a user ID parameter. This way, instructors will be able
to use their favorite survey tool and be able to properly associate
third-party survey data with edX data.