s/edX/"Stanford Online" in emails
Hardcoding it in for now until we come up with a better way to
templatize this in the emails. This probably doesn't have so much
to do with the theme as it does with the branding: perhaps we can
use some of the subdomain stuff (like we do for logos and such) to
make things work here?