Let RVM install apple-gcc42 if needed for build
With the --autolibs option set to 3, RVM will install apple-gcc42
(if needed) to build Ruby on OS X, since Ruby doesn't like to
compile with clang or the LLVM-based gcc that ships with Xcode. Now,
we don't need to manually set the `CC` environment variable to
`gcc` to get Ruby to compile.