Move Candy.js widget to its own iframe
Moving the Jabber widget to its own iframe allows us to pop it out
and in of the courseware. Refactoring more of the code (static
assets, templates, etc.) keeps the Jabber app more self-contained
so that it doesn't clutter up the main repo code.

This code needs some Jasmine tests and view tests. It also needs
some extra configuration in the settings and URLs files, namely:

* Must add jabber/templates to MAKO_TEMPLATES
* Must add jabber/static to STATICFILES_DIRS
* Must add route for ?P<course_id>/chat to lms/urls.py