Modify chat login form to not prompt for uname/pwd
After a period of inactivity, users will be disconnected from the
chat server. By default, Candy.js prompts them for a username and
password in order to reconnect. However, since we've generated both
of these for the user, they don't know what their username and
password are, and we can just reconnect them automatically.

Therefore, we don't want Candy.js to ever display a login form.
Instead, we override the `showLoginForm` function to display a
notification and a button to click when the user is ready to
reconnect to the chat server.